backend/requirements.txt
alembic==1.16.4
bidict==0.23.1
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
dnspython==2.7.0
et_xmlfile==2.0.0
Flask==3.1.1
flask-cors==6.0.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
numpy==2.2.6
openpyxl==3.1.5
pandas==2.3.2
pillow==11.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.43
tomli==2.2.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wsproto==1.2.0
gunicorn==21.2.0